refactor(graphs): remove unused coloring workflow
