+++ src/jacobian/graphs/__init__.py
@@ -1,9 +1,5 @@
-"""Graph construction, property, coloring, and composition capabilities."""
+"""Graph construction, property, and composition capabilities."""
 
-from jacobian.graphs.coloring import (
-    GraphColoringInstallation,
-    install_graph_coloring_capabilities,
-)
 from jacobian.graphs.composition import (
     GraphCompositionInstallation,
     install_graph_composition_capabilities,
@@ -15,12 +11,10 @@ from jacobian.graphs.isomorphism import (
 )
 
 __all__ = [
-    "GraphColoringInstallation",
     "GraphCompositionInstallation",
     "GraphInstallation",
     "GraphIsomorphismInstallation",
     "install_graph_capabilities",
-    "install_graph_coloring_capabilities",
     "install_graph_composition_capabilities",
     "install_graph_isomorphism",
 ]
